자가진단 상세 페이지 설문지 불러오기

+++ src/main/java/com/graduation/daily/repository/SurveyRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface SurveyRepository extends JpaRepository<Survey, Long> {
 
-    List<Survey> findByType(int type);
+    List<Survey> findByType(Long type);
 }
